Create svm_classifier.py
Spencer and I wrote a program that uses a SciKit-Learn implementation of support vector machines to classify the data. There are three classes to pick from, and we've tested two so far: LinearSVC takes about 10 minutes to run and has a success rate of 74.9%, and SVC take 30 minutes to run with an 86.7% success rate.